+++ defaults/controllers/_default/root/admin/localWorkers/view.js
@@ -1,0 +1,15 @@
+// --------------------
+// Workers Local resource controller
+// view action
+// --------------------
+
+// exports
+
+// action definition
+exports = module.exports = {
+    // functions
+
+    access: function() {
+        return false;
+    }
+};